Removal of select screen. Work on unit selection.
Selection screen was removed, as players will not be choosing their unit
layouts now. More work was done on unit selection, with the currently
selected unit implemented as well as actual unit placement and drawing
in. Will work on making units only selectable for a player on a player's
turn, and only selectable if they haven't already committed an action.
Afterward, will implement actual unit actions.